refactor: extract git automation and event helpers from orchestrator
Move emitProgress/makeEvent to src/utils/events.js and
prepareGitAutomation/finalizeGitAutomation/commitMessageFromTask to
src/git/automation.js. Orchestrator reduced from 875 to 683 lines
(-22%). Adds 11 dedicated tests. All 195 tests pass.
