+++ src/orchestrator.js
@@ -1,5 +1,4 @@
 import fs from "node:fs/promises";
-import { EventEmitter } from "node:events";
 import { createAgent } from "./agents/index.js";
 import {
   addCheckpoint,
@@ -21,37 +20,12 @@ import { runSonarScan } from "./sonar/scanner.js";
 import { shouldBlockByProfile, summarizeIssues } from "./sonar/enforcer.js";
 import { resolveRole } from "./config.js";
 import { RepeatDetector } from "./repeat-detector.js";
+import { emitProgress, makeEvent } from "./utils/events.js";
 import {
-  ensureGitRepo,
-  currentBranch,
-  fetchBase,
-  syncBaseBranch,
-  ensureBranchUpToDateWithBase,
-  createBranch,
-  buildBranchName,
-  commitAll,
-  pushBranch,
-  createPullRequest
-} from "./utils/git.js";
-
-function emitProgress(emitter, data) {
-  if (!emitter) return;
-  emitter.emit("progress", data);
-}
-
-function makeEvent(type, base, extra = {}) {
-  return {
-    type,
-    sessionId: base.sessionId,
-    iteration: base.iteration,
-    stage: base.stage,
-    status: extra.status || "ok",
-    message: extra.message || type,
-    detail: extra.detail || {},
-    elapsed: base.startedAt ? Date.now() - base.startedAt : 0,
-    timestamp: new Date().toISOString()
-  };
-}
+  commitMessageFromTask,
+  prepareGitAutomation,
+  finalizeGitAutomation
+} from "./git/automation.js";
 
 function getRepeatThreshold(config) {
   const raw =
@@ -109,84 +83,6 @@ async function runReviewerWithFallback({ reviewerName, config, logger, prompt, s
   }
 
   return { result: null, attempts };
-}
-
-function commitMessageFromTask(task) {
-  const clean = String(task || "")
-    .replace(/\s+/g, " ")
-    .trim();
-  return `feat: ${clean.slice(0, 72) || "karajan update"}`;
-}
-
-async function prepareGitAutomation({ config, task, logger, session }) {
-  const enabled = config.git.auto_commit || config.git.auto_push || config.git.auto_pr;
-  if (!enabled) return { enabled: false };
-
-  if (!(await ensureGitRepo())) {
-    throw new Error("Git automation requested but current directory is not a git repository");
-  }
-
-  const baseBranch = config.base_branch;
-  const autoRebase = config.git.auto_rebase !== false;
-  await fetchBase(baseBranch);
-
-  let branch = await currentBranch();
-  if (branch === baseBranch) {
-    await syncBaseBranch({ baseBranch, autoRebase });
-    const created = buildBranchName(config.git.branch_prefix || "feat/", task);
-    await createBranch(created);
-    branch = created;
-    logger.info(`Created working branch: ${branch}`);
-    await addCheckpoint(session, { stage: "git-prep", branch, created: true });
-  } else {
-    await ensureBranchUpToDateWithBase({ branch, baseBranch, autoRebase });
-    await addCheckpoint(session, { stage: "git-prep", branch, created: false });
-  }
-
-  return { enabled: true, branch, baseBranch, autoRebase };
-}
-
-async function finalizeGitAutomation({ config, gitCtx, task, logger, session }) {
-  if (!gitCtx?.enabled) return { git: "disabled" };
-
-  const commitMsg = config.git.commit_message || commitMessageFromTask(task);
-  let committed = false;
-  if (config.git.auto_commit) {
-    const commitResult = await commitAll(commitMsg);
-    committed = commitResult.committed;
-    await addCheckpoint(session, { stage: "git-commit", committed });
-    logger.info(committed ? "Committed changes" : "No changes to commit");
-  }
-
-  if (config.git.auto_push || config.git.auto_pr) {
-    await fetchBase(gitCtx.baseBranch);
-    await ensureBranchUpToDateWithBase({
-      branch: gitCtx.branch,
-      baseBranch: gitCtx.baseBranch,
-      autoRebase: gitCtx.autoRebase
-    });
-    await addCheckpoint(session, { stage: "git-rebase-check", branch: gitCtx.branch });
-  }
-
-  if (config.git.auto_push || config.git.auto_pr) {
-    await pushBranch(gitCtx.branch);
-    await addCheckpoint(session, { stage: "git-push", branch: gitCtx.branch });
-    logger.info(`Pushed branch: ${gitCtx.branch}`);
-  }
-
-  let prUrl = null;
-  if (config.git.auto_pr) {
-    prUrl = await createPullRequest({
-      baseBranch: gitCtx.baseBranch,
-      branch: gitCtx.branch,
-      title: commitMessageFromTask(task),
-      body: "Created by Karajan Code."
-    });
-    await addCheckpoint(session, { stage: "git-pr", branch: gitCtx.branch, pr: prUrl });
-    logger.info("Pull request created");
-  }
-
-  return { committed, branch: gitCtx.branch, prUrl };
 }
 
 export async function runFlow({ task, config, logger, flags = {}, emitter = null, askQuestion = null }) {
